Include `pdf.js`'s dependencies.
This modifies the build such that the dependencies of `pdf.js` are
included.

Installing the dependencies of this package required that we ditch
the `alpine` base image, given some of the dependencies require compiling
c / c++ code, and doing so didn't work on `alpine`. This isn't terrible,
and we could always spend more time figuring out the specific headers / libs
that `alpine` needs if we end up wanting to reduce the size of the final
iamge.