Use an optimized version of the PDF.js assets.
This modifies the UI so that we use production grade PDF.js
assets. To accomplish this I:

- Removed the submodule, in favor of committing the PDF.js
  build artifacts into the repository.
- Removed the steps we were taking in `Docker` to support
  the submodule mechanism.
- Add basic instructions for downloading the PDF.js fork
  we rely upon and executing the build.
- Remove the `pdfjs-dist` package, as it downloads a
  precompiled PDF.js distribution, in favor of
  `@types/pdfjs-dist`. This provides us with the
  typing information without downloading the prebuilt
  distribution.
- Update a few JavaScript dependencies to resolve a few
  security alerts flagged by `npm`.
- Update the build so that we properly cache from both
  build targets. This should speed up builds considerably.

In testing this locally, things seem *much* faster. The resources
are roughly 50% the size, and there's considerably less files
to download.  That said I didn't measure the precise speed up
(though we can do so if we think it'd be valuable).