Use a precompiled, optimized version of PDF.js (#7)
This modifies the repository so that the build artifacts from
PDF.js are committed to `public/`. This simplifies the build
process, and makes it easy to use an optimized version of
PDF.js.

This should provide a significant performance boost. I've also
included instructions and a script for automating the process
of updating these files.