Add a UI for viewing all papers in the system. (#8)
This adds the `/papers/` endpoint to the interface, which displays
a list of papers the system knows about.

Given it's a standalone mechanism that is likely to be discarded,
I used standard Material UI components for quickly building the
interface.

There's no pagination currently. I figured we can add that later
when it's necessary.

As part of this I had to do some acrobatics with the webpack
build. I added another HTML file and entrypoint.