Add the ability to submit feedback. (#11)
This adds the ability to submit feedback. The user can submit feedback
from the following contexts:

1. Globally
2. From the citation tooltip
3. From the drawer, after clicking a citation
4. From the symbol tooltip
5. From the drawer, after clicking a symbol

Clicking the button takes the user to a Google form, and populates
the `context` field with information from the user interface. This
mechanism allows us to decorate the feedback with metadata that helps
us examine the user's state upon preparing feedback.