Close the drawer on escape. (#16)
This adds code that closes the drawer if the escape key is pressed.
This felt like a nice convenience, and conveniently required very
little code.

I also revised the annotation event handler to use the same mechanism
for binding and removing events. We don't actively mount / remount
the top level `<ScholarReader />`, so there's no active motivator --
but in case we ever do it'll ensure things get added and removed properly.

I also had to modify `isKeypressEscape` to accept a native
`KeyboardEvent`. This worked fine as they both satisfy the expected
interface.