Add S2 actions. (#15)
This adds the "cite" and "save" actions to the paper summary
view in the drawer.  When they're clicked we kindly tell the user
they're not implemented yet, and track a Heap analytics action.

This change introduces the S2 specific citation count and
citation velocity to the UI.

I also did some retrofitting of the paper summary so that it
better matches the Semantic Scholar look / feel. This included
adjusting the border, typography and link colors.