Refactor: rename 'Props' interfaces; bind event handlers in component constructors
The binding of event handlers in component constructors is done for performance reasons.
Otherwise, as noted in the comments in ScholarReader, new event handlers will be
created at every re-render of a parent component, triggering a re-rendering of child
components. This was a major cause of unnecessary re-renders of the document (i.e.,
a change in the query field of the find bar was triggering the ScholarReader component
to re-render, which generated new event handlers for the PageOverlay, which then
triggered a very costly re-rendering of the PageOverlay object). This issue is now
fixed in this commit.