Update the health route so it's public.
This way we can use an uptime check to verify that the API is
healthy.

I removed the `livenessProbe`s, as it's something we've learned
shouldn't be used unless it's necessary.  These probes cause the
pod to be restarted when they start failing. It sounds nice in practice,
but if your server gets overwhelmed you want to give it time to
play catch-up, so that the requests being processed by the pod aren't
dropped on the floor.  The hard-restart behavior is only necessary
for pathological processes with errant behavior, like memory leaks.