Fix the paper list.
A long time ago in a galaxy far, far away...I added a list of papers
that were interesting to look at in ScholarPhi.

This PR fixes that list, in hopes that it might help with things
like https://github.com/allenai/scholar/issues/25368.

The fix required revising a database query and removing a whitelist,
which is probably ok to remove since (I think) we're beyond whatever
demo window it was added for.