Query S2's API using an API key if we have one.
This PR started as an effort to add search, but as discussed I ended
up pausing on that change in favor of eventually using S2's regular
search API. This PR instead captures changes that I made while
implementing search that seem worth preserving.

The largest functional difference is that the API now uses an API
key when querying S2's public API. This means we can safely query
for this information in the backend when preparing papers to return
for the `/papers/list` API endpoint.

To do this I did some cleanup of the way configuration parameters
are handled so that they can be accessed in a typesafe fashion.

Lastly I cleaned up the paper list UI a bit, as it was getting
a little unruly.