Retry logic and retryable exit code for Arxiv fetch failures (#244)
We're seeing a lot of "peer closed connection" errors when requesting
data from ArXiv. This may or may not be rate limiting; however we should
not treat it as an unrecoverable error.

This changeset does two things:

1) It intercepts expected failure modes in `fetch_from_arxiv`, raising a new error type
that can be caught in a driver script and used to provide a retryable exit code to
`scholarphi-pipeline`

2) Performs an in-process retry loop with more generous backoff delay.

This won't solve our woes but will allow us to at least retry sensibly.